Added ability to choose communication Mode
Added ability to choose communication Mode. The functionality of mode = 0 is still active by default and does not affect the calling of the class. The second option, mode = 1, however was added to functionality for its debugging purposes. This puts the robot into verbose mode, which provides detailed descriptions along with error codes, should they arise.